+++ lib/rearmed/apply_patches.rb
@@ -1,6 +1,3 @@
-require 'rearmed/monkey_patches/object'
-require 'rearmed/monkey_patches/string'
-require 'rearmed/monkey_patches/array'
-require 'rearmed/monkey_patches/hash'
-require 'rearmed/monkey_patches/enumerable'
-require 'rearmed/monkey_patches/date'
+Dir["rearmed/monkey_patches/*.rb"].each do |filename| 
+  require filename
+end