+++ lib/rearmed/monkey_patches/object.rb
@@ -3,7 +3,7 @@ object_enabled = Rearmed.enabled_patches[:object] == true
 Object.class_eval do
   if object_enabled || Rearmed.dig(Rearmed.enabled_patches, :object, :not_nil)
     def not_nil?
-      !nil?
+      !self.nil?
     end
   end
 
@@ -13,6 +13,24 @@ Object.class_eval do
         array = [array, *more]
       end
       array.include?(self)
+    end
+  end
+
+  if object_enabled || Rearmed.dig(Rearmed.enabled_patches, :object, :bool?)
+    def bool?
+      self.is_a?(TrueClass) || self.is_a?(FalseClass)
+    end
+  end
+
+  if object_enabled || Rearmed.dig(Rearmed.enabled_patches, :object, :false?)
+    def false?
+      self.is_a?(FalseClass)
+    end
+  end
+
+  if object_enabled || Rearmed.dig(Rearmed.enabled_patches, :object, :true?)
+    def true?
+      self.is_a?(TrueClass)
     end
   end
 end